+++ core/lib/api.web.ts
@@ -9,4 +9,7 @@ export {
     getDeviceLabelSearchMatches,
     getSupportedDeviceLabels,
 } from './devices.ts'
-export { BrowserEngineValues, type BrowserEngine } from './playwright.ts'
+export {
+    BrowserEngineValues,
+    type BrowserEngine,
+} from './playwrightBrowsers.ts'