+++ core/lib/playwright-core.d.ts
@@ -1,3 +1,4 @@
 declare module 'playwright-core/lib/server' {
     let installBrowsersForNpmInstall: (browsers: string[]) => Promise<void>
+    let registryDirectory: string
 }